Added more conditional compilation and comments to detect when cross-compilation may fail
