+++ images/runicos-base/runtime/Cargo.toml
@@ -19,7 +19,9 @@ log = "0.4.14"
 rune-runtime = { path = "../../../crates/runtime", version = "0.4.0-dev", features = ["builtins"] }
 rune-core = { path = "../../../crates/rune-core", version = "0.4.0-dev" }
 serde_json = { version = "1" }
+# Note: This doesn't cross-compile due to https://github.com/boncheolgu/tflite-rs/issues/49
 tflite = { version = "0.9.5", optional = true }
+cfg-if = "1.0.0"
 
 [features]
 default = ["tflite"]
